# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PhishGuard API",
        "description": "API untuk deteksi dan pelaporan phishing",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://be-phisguard-production.up.railway.app/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/admin/respons": {
            "post": {
                "tags": [
                    "Admin Response"
                ],
                "summary": "Create admin response for report",
                "operationId": "354f2964dd2a0255511c04c949f13165",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "report_id",
                                    "hasil_keputusan"
                                ],
                                "properties": {
                                    "report_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hasil_keputusan": {
                                        "type": "string",
                                        "example": "Confirm Valid Phishing",
                                        "enum": [
                                            "Confirm Valid Phishing",
                                            "False Positive",
                                            "Need More Info"
                                        ]
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "Mengatasnamakan Bank",
                                        "enum": [
                                            "Mengatasnamakan Bank",
                                            "E-Wallet / Fintech",
                                            "OTP / Verifikasi Akun",
                                            "Hadiah / Undian",
                                            "Paket / Kurir",
                                            "Customer Service Palsu",
                                            "Investasi Bodong",
                                            "Akun Marketplace",
                                            "Typosquatting / Domain Palsu",
                                            "Lainnya"
                                        ]
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Telah diverifikasi sebagai phishing valid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/respons": {
            "get": {
                "tags": [
                    "Response"
                ],
                "summary": "Get all responses",
                "operationId": "24ec7a2fc9cba4380758042790d22290",
                "parameters": [
                    {
                        "name": "report_id",
                        "in": "query",
                        "description": "Filter by report ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hasil_keputusan",
                        "in": "query",
                        "description": "Filter by decision result",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Confirm Valid Phishing",
                                "False Positive",
                                "Need More Info"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of responses"
                    }
                }
            }
        },
        "/respons/{id}": {
            "get": {
                "tags": [
                    "Response"
                ],
                "summary": "Get response by ID",
                "operationId": "8384e07ca652bf30fe117aa816ad26c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response detail"
                    },
                    "404": {
                        "description": "Response not found"
                    }
                }
            }
        },
        "/admin/respons/{id}": {
            "delete": {
                "tags": [
                    "Admin Response"
                ],
                "summary": "Delete response",
                "operationId": "123d9cd82dedd93b65f301f9f305c343",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response deleted successfully"
                    },
                    "404": {
                        "description": "Response not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Admin Response"
                ],
                "summary": "Update response",
                "operationId": "3fe62a68ec1e5a460ff247c334967975",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hasil_keputusan": {
                                        "type": "string",
                                        "enum": [
                                            "Confirm Valid Phishing",
                                            "False Positive",
                                            "Need More Info"
                                        ]
                                    },
                                    "kategori": {
                                        "type": "string"
                                    },
                                    "catatan": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response updated successfully"
                    },
                    "404": {
                        "description": "Response not found"
                    }
                }
            }
        }
    },
...
}